+++ main/app/repository/access.go
@@ -7,10 +7,10 @@ import (
 )
 
 type IAccessRepo interface {
-	GetPermissions(companyId uint, where ...Where) ([]model.Access, error)
-	GetAccessIds(roleIds []int, appId uint) ([]int, error)
+	GetPermissions(companyId uint64, where ...Where) ([]model.Access, error)
+	GetAccessIds(roleIds []uint64, appId uint64) ([]uint64, error)
 	CreateAccess(access *model.Access) error
-	WhereIds(accessIds []int) Where
+	WhereIds(accessIds []uint64) Where
 	WherePath(excludePath []string) Where
 	WhereIsSupplier() Where
 }
@@ -31,7 +31,7 @@ func (r *AccessRepo) CreateAccess(access *model.Access) error {
 	return r.db.Create(access).Error
 }
 
-func (r *AccessRepo) GetPermissions(companyId uint, where ...Where) ([]model.Access, error) {
+func (r *AccessRepo) GetPermissions(companyId uint64, where ...Where) ([]model.Access, error) {
 	var access []model.Access
 	for _, w := range where {
 		r.db = w(r.db)
@@ -40,13 +40,13 @@ func (r *AccessRepo) GetPermissions(companyId uint, where ...Where) ([]model.Acc
 	return access, err
 }
 
-func (r *AccessRepo) GetAccessIds(roleIds []int, appId uint) ([]int, error) {
-	var accessIds []int
+func (r *AccessRepo) GetAccessIds(roleIds []uint64, appId uint64) ([]uint64, error) {
+	var accessIds []uint64
 	err := r.db.Model(&model.RoleAccess{}).Where("role_id in (?)", roleIds).Debug().Pluck("access_id", &accessIds).Error
 	return accessIds, err
 }
 
-func (r *AccessRepo) WhereIds(accessIds []int) Where {
+func (r *AccessRepo) WhereIds(accessIds []uint64) Where {
 	return func(db *gorm.DB) *gorm.DB {
 		return db.Where("id in (?)", accessIds)
 	}
